Major refactoring: decoupling and pandas
This commit brings refactoring in two major ways:

Decoupling - Previously the codebase was one large function. Breaking it down into smaller functions (implementing or at least introducing decoupling) makes it easier to test, maintain and change.

Pandas - Introducing the use of a pandas dataframe as the target for decoding the json file allows for easier export and introduction of csv and sqlite formats, in addition to the default docx.

The test suite has been expanded reflecting the above.